Max Consecutive Ones III
The idea is to loop through each element (right index) and check if it is zero or not. If it is zero, we'll increment the zero count. At every iteration, we will check if the consecutive zeros are exceeding the allowed limit or not, once it is exceeded, we will move the window by updating the left index and the zero_count. After this, we will check for the max_length, update it and return it accordingly.